requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
akshare
pandas==2.1.3
pydantic-settings==2.1.0
langchain==0.0.338
openai==1.3.3
finrl
stable-baselines3==2.2.1
shimmy==1.3.0
